pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pokewalker-client"
version = "0.1.0"
description = "A Python client for communicating with Nintendo Pokewalker devices over USB-IrDA"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Pokewalker Client Contributors"}
]
keywords = ["pokewalker", "pokemon", "infrared", "irda", "nintendo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment",
    "Topic :: System :: Hardware",
]
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
images = [
    "Pillow>=9.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
all = [
    "pokewalker-client[images,dev]",
]

[project.scripts]
pokewalker = "pokewalker_client.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/pokewalker-client"
Documentation = "https://github.com/yourusername/pokewalker-client#readme"
Repository = "https://github.com/yourusername/pokewalker-client"

[tool.setuptools.packages.find]
where = ["."]
include = ["pokewalker_client*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line length handled by black

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
